+++ src/runtime/vue/overrides/none/Link.vue
@@ -73,7 +73,7 @@ export interface LinkSlots {
 </script>
 
 <script setup lang="ts">
-import { computed } from 'vue'
+import { computed, inject } from 'vue'
 import { defu } from 'defu'
 import { hasProtocol } from 'ufo'
 import { useAppConfig } from '#imports'
@@ -162,6 +162,18 @@ const linkRel = computed(() => {
 
   return null
 })
+
+const handleNavigation = inject<((event: MouseEvent, context: { href: string, external: boolean, target?: string | null }) => void) | undefined>('bitrix24ui:router', undefined)
+
+const navigate = handleNavigation
+  ? (e: MouseEvent) => {
+      handleNavigation(e, {
+        href: href.value || '',
+        external: isExternal.value,
+        target: props.target || (isExternal.value ? '_blank' : undefined)
+      })
+    }
+  : undefined
 </script>
 
 <template>
@@ -172,8 +184,8 @@ const linkRel = computed(() => {
         as,
         type,
         disabled,
-        href: href,
-        navigate: undefined,
+        href,
+        navigate,
         rel: linkRel,
         target: target || (isExternal ? '_blank' : undefined),
         isExternal,
@@ -188,8 +200,8 @@ const linkRel = computed(() => {
       as,
       type,
       disabled,
-      href: href,
-      navigate: undefined,
+      href,
+      navigate,
       rel: linkRel,
       target: target || (isExternal ? '_blank' : undefined),
       isExternal