+++ src/runtime/vue/plugins/router.ts
@@ -1,0 +1,9 @@
+import type { Plugin } from 'vue'
+
+export default {
+  install(app, options) {
+    if (options?.router && typeof options.router === 'function') {
+      app.provide('bitrix24ui:router', options.router)
+    }
+  }
+} satisfies Plugin