+++ CHANGELOG.md
@@ -62,6 +62,7 @@
 * fix(Banner): isolate banner visibility using per-instance CSS variables
 * chore(deps): update dependency unplugin-auto-import to v21
 * fix(defineShortcuts): check shift modifier for special character shortcuts
+* feat(Link): support custom navigate function in vue
 
 ## [2.2.1](https://github.com/bitrix24/b24ui/compare/v2.1.17...v2.2.1) (2025-12-18)